feat(flask): add API for fetching status of jobs
